Add files via upload
The project includes reading data on two .csv files which one of them included selected part of gene's IDs, and the other consisted of the whole genes with their names, IDs, and the measured log2FC values. This is the first time I've worked on both biology and data science.